Move DataDocument and LogLine out of the database layer
This is the last edge that kept the compiler reaching the database. Any module's
Python code imports inmanta.agent.handler, which annotated and built
data.LogLine, so importing the handler API loaded inmanta/data/__init__.py and
with it sqlalchemy.

LogLine subclasses DataDocument, which data/__init__.py documents as explicitly
not stored in the database: it exists to give the RPC something it can serialize.
Neither belongs in the access layer, and DataDocument had only two references
there. Both move to inmanta.data_document, which needs nothing above the
foundation. inmanta.data re-exports them, so data.LogLine keeps working for the
server and deploy code that uses it.

With this and the DTO move, a compile no longer imports the database at all:

  import inmanta.app          1739 -> 1530 modules, inmanta.data and sqlalchemy absent
  real compile, std master    1744 -> 1535 modules, likewise absent

The four baselined typing errors for these two classes follow them to
src/inmanta/data_document.py; they are unchanged.
