+++ mypy-baseline.txt
@@ -310,15 +310,12 @@ src/inmanta/data/__init__.py:0: error: Incompatible return value type (got "list
 src/inmanta/data/__init__.py:0: error: Incompatible return value type (got "list[dict[str, object]]", expected "list[ConfigurationModel]")  [return-value]
 src/inmanta/data/__init__.py:0: error: Incompatible return value type (got "list[object]", expected "list[str]")  [return-value]
 src/inmanta/data/__init__.py:0: error: Incompatible return value type (got "object", expected "T")  [return-value]
-src/inmanta/data/__init__.py:0: error: Incompatible return value type (got "object", expected "list[Any]")  [return-value]
-src/inmanta/data/__init__.py:0: error: Incompatible return value type (got "object", expected "str")  [return-value]
 src/inmanta/data/__init__.py:0: error: Incompatible types in "await" (actual type "AbstractAsyncContextManager[Connection, bool | None]", expected type "Awaitable[Any]")  [misc]
 src/inmanta/data/__init__.py:0: error: Incompatible types in "await" (actual type "Pool", expected type "Awaitable[Any]")  [misc]
 src/inmanta/data/__init__.py:0: error: Incompatible types in assignment (expression has type "DocumentMeta", variable has type "type[BaseDocument]")  [assignment]
 src/inmanta/data/__init__.py:0: error: Incompatible types in assignment (expression has type "dict[str, object]", variable has type "Record")  [assignment]
 src/inmanta/data/__init__.py:0: error: Incompatible types in assignment (expression has type "object", variable has type "int")  [assignment]
 src/inmanta/data/__init__.py:0: error: Incompatible types in assignment (expression has type "object", variable has type "int")  [assignment]
-src/inmanta/data/__init__.py:0: error: Incompatible types in assignment (expression has type "object", variable has type "str")  [assignment]
 src/inmanta/data/__init__.py:0: error: Incompatible types in assignment (expression has type "str", variable has type "ResourceIdStr")  [assignment]
 src/inmanta/data/__init__.py:0: error: Incompatible types in assignment (expression has type "str", variable has type "ResourceIdStr")  [assignment]
 src/inmanta/data/__init__.py:0: error: Incompatible types in assignment (expression has type "str", variable has type "dict[str, object]")  [assignment]
@@ -337,7 +334,6 @@ src/inmanta/data/__init__.py:0: error: Missing type arguments for generic type "
 src/inmanta/data/__init__.py:0: error: Missing type arguments for generic type "Field"  [type-arg]
 src/inmanta/data/__init__.py:0: error: Missing type arguments for generic type "Field"  [type-arg]
 src/inmanta/data/__init__.py:0: error: Missing type arguments for generic type "dict"  [type-arg]
-src/inmanta/data/__init__.py:0: error: Missing type arguments for generic type "list"  [type-arg]
 src/inmanta/data/__init__.py:0: error: Need type annotation for "_updates" (hint: "_updates: dict[<type>, <type>] = ...")  [var-annotated]
 src/inmanta/data/__init__.py:0: error: No overload variant of "__add__" of "list" matches argument type "Sequence[tuple[str, ColumnType, PagingOrder]]"  [operator]
 src/inmanta/data/__init__.py:0: error: No overload variant of "int" matches argument type "object"  [call-overload]
@@ -466,18 +462,16 @@ src/inmanta/data/dataview.py:0: error: Argument 1 to "parse_id" of "Id" has inco
 src/inmanta/data/dataview.py:0: error: Return type "Sequence[dict[str, str]]" of "construct_dtos" incompatible with return type "Sequence[DiscoveredResourceOutput]" in supertype "DataView"  [override]
 src/inmanta/data/dataview.py:0: error: Unsupported left operand type for + ("object")  [operator]
 src/inmanta/data/dataview.py:0: error: Unsupported left operand type for + ("object")  [operator]
-src/inmanta/dto.py:0: error: Argument 1 of "__setstate__" is incompatible with supertype "pydantic.main.BaseModel"; supertype defines the argument type as "dict[Any, Any]"  [override]
-src/inmanta/dto.py:0: error: Function is missing a type annotation  [no-untyped-def]
-src/inmanta/dto.py:0: error: Return type "str" of "__getstate__" incompatible with return type "dict[Any, Any]" in supertype "pydantic.main.BaseModel"  [override]
-src/inmanta/dto.py:0: error: Unused "type: ignore" comment, use narrower [prop-decorator] instead of [misc] code  [unused-ignore]
-src/inmanta/dto.py:0: note: See https://mypy.readthedocs.io/en/stable/common_issues.html#incompatible-overrides
-src/inmanta/dto.py:0: note: This violates the Liskov substitution principle
 src/inmanta/data/schema.py:0: error: No overload variant of "set" matches argument type "object"  [call-overload]
 src/inmanta/data/schema.py:0: error: Returning Any from function declared to return "bool"  [no-any-return]
 src/inmanta/data/schema.py:0: error: Returning Any from function declared to return "set[int]"  [no-any-return]
 src/inmanta/data/schema.py:0: note: Possible overload variants:
 src/inmanta/data/schema.py:0: note: def [_T] set() -> set[_T]
 src/inmanta/data/schema.py:0: note: def [_T] set(Iterable[_T], /) -> set[_T]
+src/inmanta/data_document.py:0: error: Incompatible return value type (got "object", expected "list[Any]")  [return-value]
+src/inmanta/data_document.py:0: error: Incompatible return value type (got "object", expected "str")  [return-value]
+src/inmanta/data_document.py:0: error: Incompatible types in assignment (expression has type "object", variable has type "str")  [assignment]
+src/inmanta/data_document.py:0: error: Missing type arguments for generic type "list"  [type-arg]
 src/inmanta/deploy/scheduler.py:0: error: "object" has no attribute "__iter__"; maybe "__dir__" or "__str__"? (not iterable)  [attr-defined]
 src/inmanta/deploy/state.py:0: error: "object" has no attribute "__iter__"; maybe "__dir__" or "__str__"? (not iterable)  [attr-defined]
 src/inmanta/deploy/state.py:0: error: Argument "attribute_hash" to "ResourceIntent" has incompatible type "object"; expected "str"  [arg-type]
@@ -490,6 +484,12 @@ src/inmanta/deploy/state.py:0: error: Unsupported left operand type for > ("obje
 src/inmanta/docstring_parser.py:0: error: Argument 1 to "DocString" has incompatible type "str"; expected "Docstring"  [arg-type]
 src/inmanta/docstring_parser.py:0: error: Incompatible types in assignment (expression has type "Docstring", variable has type "str")  [assignment]
 src/inmanta/docstring_parser.py:0: error: Value expression in dictionary comprehension has incompatible type "str | None"; expected type "str"  [misc]
+src/inmanta/dto.py:0: error: Argument 1 of "__setstate__" is incompatible with supertype "pydantic.main.BaseModel"; supertype defines the argument type as "dict[Any, Any]"  [override]
+src/inmanta/dto.py:0: error: Function is missing a type annotation  [no-untyped-def]
+src/inmanta/dto.py:0: error: Return type "str" of "__getstate__" incompatible with return type "dict[Any, Any]" in supertype "pydantic.main.BaseModel"  [override]
+src/inmanta/dto.py:0: error: Unused "type: ignore" comment, use narrower [prop-decorator] instead of [misc] code  [unused-ignore]
+src/inmanta/dto.py:0: note: See https://mypy.readthedocs.io/en/stable/common_issues.html#incompatible-overrides
+src/inmanta/dto.py:0: note: This violates the Liskov substitution principle
 src/inmanta/env.py:0: error: Argument 1 to "parse_requirements" has incompatible type "Sequence[CanonicalRequirement]"; expected "Sequence[str]"  [arg-type]
 src/inmanta/env.py:0: error: Incompatible return value type (got "tuple[str | None, Loader | None] | None", expected "tuple[str | None, Loader] | None")  [return-value]
 src/inmanta/env.py:0: error: Item "None" of "ModuleSpec | None" has no attribute "submodule_search_locations"  [union-attr]
