+++ src/inmanta/agent/handler.py
@@ -32,9 +32,10 @@ from typing import Any, Callable, Generic, Optional, TypeVar, Union, cast, overl
 from tornado import concurrent
 
 import inmanta
-from inmanta import const, data, protocol, resources, tracing
+from inmanta import const, protocol, resources, tracing
 from inmanta.agent.cache import AgentCache
 from inmanta.const import ParameterSource, ResourceState
+from inmanta.data_document import LogLine
 from inmanta.dto import AttributeStateChange, BaseModel, DiscoveredResourceInput
 from inmanta.protocol import Result, json_encode
 from inmanta.stable_api import stable_api
@@ -287,7 +288,7 @@ class HandlerContext(LoggerABC):
         self._action_id = action_id
         self._status: Optional[ResourceState] = None
         self._resource_state: Optional[const.HandlerResourceState] = None
-        self._logs: list[data.LogLine] = []
+        self._logs: list[LogLine] = []
         self.logger: logging.Logger
         if logger is None:
             self.logger = LOGGER
@@ -331,7 +332,7 @@ class HandlerContext(LoggerABC):
         return self._resource_state
 
     @property
-    def logs(self) -> list[data.LogLine]:
+    def logs(self) -> list[LogLine]:
         return self._logs
 
     def set_status(self, status: const.ResourceState) -> None:
@@ -514,7 +515,7 @@ class HandlerContext(LoggerABC):
 
         packaged_kwargs = {k: clean_arg_value(k, v) for k, v in kwargs.items()}
 
-        log = data.LogLine.log(level, msg, **packaged_kwargs)
+        log = LogLine.log(level, msg, **packaged_kwargs)
         self.logger.log(level, "resource %s: %s", self._resource.id.resource_version_str(), log._data["msg"], exc_info=exc_info)
         self._logs.append(log)
